# docs/technical/openapi.yaml
openapi: 3.1.0
info:
  title: Horror Story Research API
  description: |
    Local-only API server for research card management and deduplication signal evaluation.

    ## Features
    - **Research Cards**: Generate horror research cards using local Ollama LLM
    - **Validation**: Validate existing research card quality
    - **Listing**: Browse and filter research cards
    - **Dedup Signals**: Evaluate story similarity (LOW/MEDIUM/HIGH)
    - **Resource Management**: Monitor Ollama model lifecycle

    ## Phase B+ Additions
    - FAISS-based semantic deduplication
    - Story seed generation and registry
    - Ollama resource auto-cleanup

    ## Usage
    ```bash
    # Start server
    uvicorn src.api.main:app --host 127.0.0.1 --port 8000
    ```

    ## Authentication (Optional)
    API key authentication is disabled by default. To enable:
    ```bash
    export API_AUTH_ENABLED=true
    export API_KEY=your-secure-key
    ```
    When enabled, include `X-API-Key` header in all requests (except `/health` and `/resource/status`).

    ## Note
    This API is designed for **local use only**. All operations connect to local Ollama instance.
  version: 1.6.0
  contact:
    name: VinylStage
    email: mcpe9869@gmail.com
  license:
    name: Private
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /research/run:
    post:
      tags:
        - research
      summary: Execute research generation
      description: |
        Triggers research generation via Ollama LLM.

        This endpoint spawns a subprocess to execute the research_executor CLI.
        The research card is generated and saved to disk.

        **Note**: This operation may take 30-120 seconds depending on the model.
      operationId: runResearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchRunRequest'
            examples:
              basic:
                summary: Basic research request
                value:
                  topic: "한국 아파트 공포"
                  tags: ["apartment", "korean"]
              withModel:
                summary: With model override
                value:
                  topic: "지하철 도시전설"
                  tags: ["subway", "urban-legend"]
                  model: "qwen3:30b"
                  timeout: 300
      responses:
        '200':
          description: Research execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchRunResponse'
              examples:
                success:
                  summary: Successful generation
                  value:
                    card_id: "RC-20260111-123456"
                    status: "complete"
                    message: null
                    output_path: "data/research/2026/01/RC-20260111-123456.json"
                timeout:
                  summary: Timeout error
                  value:
                    card_id: ""
                    status: "timeout"
                    message: "Research execution timed out after 300s"
                    output_path: null
  /research/validate:
    post:
      tags:
        - research
      summary: Validate research card
      description: |
        Validates an existing research card for quality and completeness.

        Quality scores:
        - **good**: All required fields present and valid
        - **partial**: Some fields missing or incomplete
        - **incomplete**: Major fields missing
      operationId: validateResearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchValidateRequest'
            example:
              card_id: "RC-20260111-123456"
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchValidateResponse'
              examples:
                valid:
                  summary: Valid card
                  value:
                    card_id: "RC-20260111-123456"
                    is_valid: true
                    quality_score: "good"
                    message: "Validation passed"
                notFound:
                  summary: Card not found
                  value:
                    card_id: "RC-20260111-999999"
                    is_valid: false
                    quality_score: "not_found"
                    message: "Card not found: data/research/2026/01/RC-20260111-999999.json"
  /research/list:
    get:
      tags:
        - research
      summary: List research cards
      description: |
        Lists research cards with optional filtering and pagination.

        Cards are sorted by creation date (newest first).
      operationId: listResearch
      parameters:
        - name: limit
          in: query
          description: Maximum number of cards to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: quality
          in: query
          description: Filter by quality score
          schema:
            type: string
            enum:
              - good
              - partial
              - incomplete
      responses:
        '200':
          description: List of research cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchListResponse'
  /research/dedup:
    post:
      tags:
        - research
      summary: Check research card for semantic duplicates
      description: |
        Checks a research card against existing cards using FAISS semantic similarity.

        **Similarity Thresholds:**
        - **LOW** (< 0.70): Unique content
        - **MEDIUM** (0.70-0.85): Some overlap
        - **HIGH** (≥ 0.85): High similarity (potential duplicate)

        Uses nomic-embed-text model via Ollama (768 dimensions).
      operationId: checkResearchDedup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchDedupCheckRequest'
            example:
              card_id: "RC-20260113-120000"
      responses:
        '200':
          description: Dedup check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchDedupCheckResponse'
              examples:
                low:
                  summary: Low similarity (unique)
                  value:
                    card_id: "RC-20260113-120000"
                    signal: "LOW"
                    similarity_score: 0.45
                    nearest_card_id: "RC-20260112-090000"
                    similar_cards:
                      - card_id: "RC-20260112-090000"
                        similarity_score: 0.45
                        title: "Urban Horror Research"
                    index_size: 52
                    message: "Card is sufficiently unique"
                high:
                  summary: High similarity (duplicate)
                  value:
                    card_id: "RC-20260113-120000"
                    signal: "HIGH"
                    similarity_score: 0.92
                    nearest_card_id: "RC-20260112-143000"
                    similar_cards:
                      - card_id: "RC-20260112-143000"
                        similarity_score: 0.92
                        title: "Korean Apartment Horror"
                    index_size: 52
                    message: "High similarity detected"
  /research/matching-templates:
    post:
      tags:
        - research
      summary: Find matching templates for research card
      description: |
        Finds story templates that match a research card's canonical affinity.

        **Cross-pipeline CK matching** (Issue #21):
        - Research cards have canonical_affinity (extracted from content)
        - Templates have canonical_core (defined in template config)
        - This endpoint matches them to suggest suitable templates

        Returns templates sorted by match score (highest first).
      operationId: matchResearchTemplates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchMatchingTemplatesRequest'
            examples:
              default:
                summary: Default matching
                value:
                  card_id: "RC-20260115-143052"
                  max_templates: 5
                  min_score: 0.5
              strict:
                summary: Strict matching (higher threshold)
                value:
                  card_id: "RC-20260115-143052"
                  max_templates: 3
                  min_score: 0.7
      responses:
        '200':
          description: Matching templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchMatchingTemplatesResponse'
        '404':
          description: Research card not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Research card not found: RC-20260115-143052"
...
components:
  schemas:
...
    ResearchRunRequest:
      type: object
      required:
        - topic
      properties:
        topic:
          type: string
          description: Research topic to analyze
          minLength: 5
          maxLength: 500
          example: "한국 아파트 공포"
        tags:
          type: array
          items:
            type: string
          description: Optional tags for categorization
          default: []
          example: ["apartment", "korean"]
        model:
          type: string
          description: Ollama model override
          example: "qwen3:30b"
        timeout:
          type: integer
          description: Timeout in seconds
          minimum: 30
          maximum: 600
          default: 300
    ResearchRunResponse:
      type: object
      required:
        - card_id
        - status
      properties:
        card_id:
          type: string
          description: Generated card ID (empty on error)
          example: "RC-20260111-123456"
        status:
          type: string
          description: Execution status
          enum:
            - complete
            - error
            - timeout
        message:
          type: string
          nullable: true
          description: Status message
        output_path:
          type: string
          nullable: true
          description: Path to output file
    ResearchValidateRequest:
      type: object
      required:
        - card_id
      properties:
        card_id:
          type: string
          description: Card ID to validate
          pattern: "^RC-\\d{8}-\\d{6}$"
          example: "RC-20260111-123456"
    ResearchValidateResponse:
      type: object
      required:
        - card_id
        - is_valid
        - quality_score
      properties:
        card_id:
          type: string
          description: Validated card ID
        is_valid:
          type: boolean
          description: Whether card is valid
        quality_score:
          type: string
          description: Quality score
          enum:
            - good
            - partial
            - incomplete
            - not_found
            - invalid_id
            - error
        message:
          type: string
          nullable: true
          description: Validation details
    ResearchCardSummary:
      type: object
      required:
        - card_id
        - title
        - topic
        - quality_score
        - created_at
      properties:
        card_id:
          type: string
          description: Card identifier
          example: "RC-20260111-123456"
        title:
          type: string
          description: Research card title
          example: "아파트 엘리베이터의 공포"
        topic:
          type: string
          description: Original research topic
        quality_score:
          type: string
          description: Quality assessment
        created_at:
          type: string
          description: Creation date (YYYY-MM-DD)
          example: "2026-01-11"
    ResearchListResponse:
      type: object
      required:
        - cards
        - total
        - limit
        - offset
      properties:
        cards:
          type: array
          items:
            $ref: '#/components/schemas/ResearchCardSummary'
          description: List of research cards
        total:
          type: integer
          description: Total number of cards returned
        limit:
          type: integer
          description: Requested limit
        offset:
          type: integer
          description: Requested offset
        message:
          type: string
          nullable: true
          description: Status message
...
    ResearchDedupCheckRequest:
      type: object
      required:
        - card_id
      properties:
        card_id:
          type: string
          description: Card ID to check for duplicates
          pattern: "^RC-\\d{8}-\\d{6}$"
    SimilarCard:
      type: object
      required:
        - card_id
        - similarity_score
      properties:
        card_id:
          type: string
          description: Similar card identifier
        similarity_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Similarity score (0.0-1.0)
        title:
          type: string
          nullable: true
          description: Card title
    ResearchDedupCheckResponse:
      type: object
      required:
        - card_id
        - signal
        - similarity_score
        - index_size
      properties:
        card_id:
          type: string
          description: Checked card ID
        signal:
          type: string
          description: Dedup signal level
          enum:
            - LOW
            - MEDIUM
            - HIGH
        similarity_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Highest similarity score (0.0-1.0)
        nearest_card_id:
          type: string
          nullable: true
          description: Most similar card ID
        similar_cards:
          type: array
          items:
            $ref: '#/components/schemas/SimilarCard'
          default: []
          description: List of similar cards
        index_size:
          type: integer
          default: 0
          description: Number of cards in FAISS index
        message:
          type: string
          nullable: true
          description: Status message
    MatchingTemplateItem:
      type: object
      required:
        - template_id
        - template_name
        - match_score
      properties:
        template_id:
          type: string
          description: Template ID (e.g., T-SYS-001)
        template_name:
          type: string
          description: Human-readable template name
        match_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Match score (0.0-1.0)
        canonical_core:
          type: object
          additionalProperties: true
          default: {}
          description: Template's canonical core values
        match_details:
          type: object
          nullable: true
          additionalProperties: true
          description: Per-dimension match breakdown
    ResearchMatchingTemplatesRequest:
      type: object
      required:
        - card_id
      properties:
        card_id:
          type: string
          description: Research card ID to match against templates
          pattern: "^RC-\\d{8}-\\d{6}$"
        max_templates:
          type: integer
          minimum: 1
          maximum: 15
          default: 5
          description: Maximum number of templates to return
        min_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          default: 0.5
          description: Minimum match score threshold (0.0-1.0)
    ResearchMatchingTemplatesResponse:
      type: object
      required:
        - card_id
        - matching_templates
        - total_templates
      properties:
        card_id:
          type: string
          description: Research card ID that was matched
        matching_templates:
          type: array
          items:
            $ref: '#/components/schemas/MatchingTemplateItem'
          default: []
          description: List of matching templates sorted by score (highest first)
        total_templates:
          type: integer
          description: Total number of templates evaluated
        card_affinity:
          type: object
          nullable: true
          additionalProperties: true
          description: The research card's canonical affinity used for matching
        message:
          type: string
          nullable: true
          description: Status message

    # FastAPI standard validation error schemas
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication (optional).
        Required when API_AUTH_ENABLED=true.
        Excluded endpoints: /health, /resource/status
security:
  - ApiKeyAuth: []
  - {}  # Also allow unauthenticated access (when API_AUTH_ENABLED=false)